[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -2,10 +2,10 @@
 
 int main(void) {
 
-  int q, w;
-  scanf("%d %d", &q, &w);
+  long int q, w;
+  scanf("%ld %ld", &q, &w);
 
-  printf("%d", (q+w)*(q-w));
+  printf("%ld", (q+w)*(q-w));
   
   return 0;
 }
